refactor(a11y): delete Tooltip rather than fix its 1.4.13 failure (bpdiscord-lye)
Tooltip was 155 lines of hover/focus state machine, fixed positioning, scroll
re-anchoring and Esc dismissal, serving two consumers, with an open WCAG 1.4.13
failure: the panel is pointer-events-none, so the pointer cannot travel onto it
without the panel disappearing.

Reading the two consumers settled it.

The histogram's ten bar tooltips lived inside a container already marked
aria-hidden="true" — its comment says the bars are decorative and the sr-only
table is the text alternative. Those tooltips reached no assistive technology
at all. They were a hover label for sighted mouse users, costing a widget with
a conformance gap. The sr-only table is untouched, so screen reader users lose
nothing. Sighted mouse users lose the per-bar counts; see the ticket for the
disclosure follow-up if that turns out to matter.

TasteCompatibility's info tooltip becomes CollapsibleSection, which the repo
already uses for exactly this and which is already remediated. Click to open
rather than hover means 1.4.13 does not apply at all, and the explanation is
now reachable by keyboard, which it never was.

The e2e harness survives — it is the only place in this repo that tests real
focus and layout, and bpdiscord-dsp points it at Modal next. Four of its five
specs were tooltip positioning and go with the component; the fifth proves the
histogram adds no tab stops and is kept as histogram.spec.ts.
