+++ src/client/e2e/harness/harness.tsx
@@ -1,5 +1,4 @@
 import { createRoot } from "react-dom/client";
-import Tooltip from "../../components/Tooltip";
 import RatingDistributionHistogram from "../../components/RatingDistributionHistogram";
 import "../../index.css";
 
@@ -9,8 +8,8 @@ const DISTRIBUTION = [
   { rating: 4.5, count: 12 },
 ];
 
-// The spacer pushes the trigger below the fold so Tab has to scroll it into
-// view — the condition under which the popup used to mis-anchor.
+// The spacer keeps the histogram below the fold, so a Tab sweep has to travel
+// past it to reach the button after it.
 const Harness = () => (
   <main style={{ padding: 24 }}>
     <button type="button" data-testid="before">
@@ -19,19 +18,15 @@ const Harness = () => (
 
     <div data-testid="spacer" style={{ height: "150vh" }} />
 
-    <div style={{ textAlign: "center" }}>
-      <Tooltip content="what this means">
-        <button type="button" data-testid="trigger">
-          info
-        </button>
-      </Tooltip>
-    </div>
-
     <div data-testid="histogram" style={{ marginTop: 48 }}>
       <RatingDistributionHistogram distribution={DISTRIBUTION} />
     </div>
 
     <div style={{ height: "150vh" }} />
+
+    <button type="button" data-testid="after">
+      after
+    </button>
   </main>
 );
